third_party/blink/web_tests/platform/generic/fast/xmlhttprequest/xmlhttprequest-responsetype-text-expected.txt

Tests XMLHttpRequest 'text' loading with the .responseType and .response attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

responseText
<?xml version="1.0"?>
<!DOCTYPE doc [
<!ATTLIST d id ID #IMPLIED>
]>
<doc>
  <foo xmlns="foobar">One</foo> <x:bar xmlns:x="barfoo">Two</x:bar>
  <d id="id3">Three</d>
</doc>



PASS responseType property exists.
PASS response property exists.
PASS xhr.responseType has been correctly set to 'text'.
PASS DONE LOADING
PASS received response object of type : string.
PASS exception correctly thrown when xhr.responseType is set to valid value too late in the loading process : InvalidStateError: Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be set if the object's state is LOADING or DONE..
PASS xhr.response == xhr.responseText.
PASS successfullyParsed is true

TEST COMPLETE
